Consolidated re-usable functions to a separate go file
